requirements.txt
infi.systray

# Building requires:
PyInstall